chore: move staging blocks' shadow code to shadow.rs
