+++ stackslib/src/chainstate/nakamoto/staging_blocks.rs
@@ -273,23 +273,6 @@ impl<'a> NakamotoStagingBlocksConnRef<'a> {
         Ok(res.is_some())
     }
 
-    /// Determine if we have a particular block with the given index hash.
-    /// Returns Ok(true) if so
-    /// Returns Ok(false) if not
-    /// Returns Err(..) on DB error
-    pub fn has_shadow_nakamoto_block_with_index_hash(
-        &self,
-        index_block_hash: &StacksBlockId,
-    ) -> Result<bool, ChainstateError> {
-        let qry = "SELECT 1 FROM nakamoto_staging_blocks WHERE index_block_hash = ?1 AND obtain_method = ?2";
-        let args = params![
-            index_block_hash,
-            &NakamotoBlockObtainMethod::Shadow.to_string()
-        ];
-        let res: Option<i64> = query_row(self, qry, args)?;
-        Ok(res.is_some())
-    }
-
     /// Get the block ID, processed-status, orphan-status, and signing weight of the non-orphaned
     /// block with the given consensus hash and sighash with the most amount of signatures.
     /// There will be at most one such block.
@@ -526,49 +509,6 @@ impl<'a> NakamotoStagingBlocksConnRef<'a> {
         let res: Option<u64> = query_row(self, sql, args)?;
         Ok(res.is_some())
     }
-
-    /// Is this a shadow tenure?
-    /// If any block is a shadow block in the tenure, they must all be.
-    ///
-    /// Returns true if the tenure has at least one shadow block.
-    pub fn is_shadow_tenure(
-        &self,
-        consensus_hash: &ConsensusHash,
-    ) -> Result<bool, ChainstateError> {
-        let qry = "SELECT 1 FROM nakamoto_staging_blocks WHERE consensus_hash = ?1 AND obtain_method = ?2";
-        let args = rusqlite::params![
-            consensus_hash,
-            NakamotoBlockObtainMethod::Shadow.to_string()
-        ];
-        let present: Option<u32> = query_row(self, qry, args)?;
-        Ok(present.is_some())
-    }
-
-    /// Shadow blocks, unlike Stacks blocks, have a unique place in the chain history.
-    /// They are inserted post-hoc, so they and their underlying burnchain blocks don't get
-    /// invalidated via a fork.  A consensus hash can identify (1) no tenures, (2) a single
-    /// shadow tenure, or (3) one or more non-shadow tenures.
-    ///
-    /// This is important when downloading a tenure that is ended by a shadow block, since it won't
-    /// be processed beforehand and its hash isn't learned from the burnchain (so we must be able
-    /// to infer that if this is a shadow tenure, none of the blocks in it have siblings).
-    pub fn get_shadow_tenure_start_block(
-        &self,
-        ch: &ConsensusHash,
-    ) -> Result<Option<NakamotoBlock>, ChainstateError> {
-        let qry = "SELECT data FROM nakamoto_staging_blocks WHERE consensus_hash = ?1 AND obtain_method = ?2 ORDER BY height DESC LIMIT 1";
-        let args = params![ch, &NakamotoBlockObtainMethod::Shadow.to_string()];
-        let res: Option<Vec<u8>> = query_row(self, qry, args)?;
-        let Some(block_bytes) = res else {
-            return Ok(None);
-        };
-        let block = NakamotoBlock::consensus_deserialize(&mut block_bytes.as_slice())?;
-        if !block.is_shadow_block() {
-            error!("Staging DB corruption: expected shadow block from {}", ch);
-            return Err(DBError::Corruption.into());
-        }
-        Ok(Some(block))
-    }
 }
 
 impl<'a> NakamotoStagingBlocksTx<'a> {
@@ -647,7 +587,7 @@ impl<'a> NakamotoStagingBlocksTx<'a> {
             obtain_method
         };
 
-        if self.is_shadow_tenure(&block.header.consensus_hash)? && !block.is_shadow_block() {
+        if self.conn().is_shadow_tenure(&block.header.consensus_hash)? && !block.is_shadow_block() {
             return Err(ChainstateError::InvalidStacksBlock(
                 "Tried to insert a non-shadow block into a shadow tenure".into(),
             ));
@@ -745,77 +685,6 @@ impl<'a> NakamotoStagingBlocksTx<'a> {
         Ok(())
     }
 
-    /// Is this a shadow tenure?
-    pub fn is_shadow_tenure(
-        &self,
-        consensus_hash: &ConsensusHash,
-    ) -> Result<bool, ChainstateError> {
-        let qry = "SELECT 1 FROM nakamoto_staging_blocks WHERE consensus_hash = ?1 AND obtain_method = ?2";
-        let args = rusqlite::params![
-            consensus_hash,
-            &NakamotoBlockObtainMethod::Shadow.to_string()
-        ];
-        let present: Option<u32> = query_row(self, qry, args)?;
-        Ok(present.is_some())
-    }
-
-    /// Add a shadow block.
-    /// Fails if there are any non-shadow blocks present in the tenure.
-    pub fn add_shadow_block(&self, shadow_block: &NakamotoBlock) -> Result<(), ChainstateError> {
-        if !shadow_block.is_shadow_block() {
-            return Err(ChainstateError::InvalidStacksBlock(
-                "Not a shadow block".into(),
-            ));
-        }
-        let block_id = shadow_block.block_id();
-
-        // is this block stored already?
-        let qry = "SELECT 1 FROM nakamoto_staging_blocks WHERE index_block_hash = ?1";
-        let args = params![block_id];
-        let present: Option<i64> = query_row(self, qry, args)?;
-        if present.is_some() {
-            return Ok(());
-        }
-
-        // this tenure must be empty, or it must be a shadow tenure
-        let qry = "SELECT 1 FROM nakamoto_staging_blocks WHERE consensus_hash = ?1";
-        let args = rusqlite::params![&shadow_block.header.consensus_hash];
-        let present: Option<u32> = query_row(self, qry, args)?;
-        if present.is_some() && !self.is_shadow_tenure(&shadow_block.header.consensus_hash)? {
-            return Err(ChainstateError::InvalidStacksBlock(
-                "Shadow block cannot be inserted into non-empty non-shadow tenure".into(),
-            ));
-        }
-
-        // there must not be a block at this height in this tenure
-        let qry = "SELECT 1 FROM nakamoto_staging_blocks WHERE consensus_hash = ?1 AND height = ?2";
-        let args = rusqlite::params![
-            &shadow_block.header.consensus_hash,
-            u64_to_sql(shadow_block.header.chain_length)?
-        ];
-        let present: Option<u32> = query_row(self, qry, args)?;
-        if present.is_some() {
-            return Err(ChainstateError::InvalidStacksBlock(format!(
-                "Conflicting block at height {} in tenure {}",
-                shadow_block.header.chain_length, &shadow_block.header.consensus_hash
-            )));
-        }
-
-        // the shadow block is crafted post-hoc, so we know the consensus hash exists.
-        // thus, it's always burn-attachable
-        let burn_attachable = true;
-
-        // shadow blocks cannot be replaced
-        let signing_weight = u32::MAX;
-
-        self.store_block(
-            shadow_block,
-            burn_attachable,
-            signing_weight,
-            NakamotoBlockObtainMethod::Shadow,
-        )?;
-        Ok(())
-    }
 }
 
 impl StacksChainState {
